added swagger + docs, reorganized endpoints:users, projects into routes folder.
