+++ api/package.json
@@ -15,6 +15,8 @@
     "express": "^4.18.2",
     "knex": "^2.5.1",
     "nodemon": "^3.0.1",
-    "pg": "^8.11.3"
+    "pg": "^8.11.3",
+    "swagger-jsdoc": "^6.2.8",
+    "swagger-ui-express": "^5.0.1"
   }
 }
